fix: downgrade `fuel-core-client` to `v0.17.2` (#1028)
* Downgrade fuel-core-client to 0.17.2; indexers need to be fixed

* Fix indexers; just two more errors to go

* I think it works?

* Fix tests

+++ packages/fuel-indexer-database/postgres/src/lib.rs
@@ -635,18 +635,18 @@ pub async fn last_block_height_for_indexer(
     conn: &mut PoolConnection<Postgres>,
     namespace: &str,
     identifier: &str,
-) -> sqlx::Result<u32> {
+) -> sqlx::Result<u64> {
     let query = format!(
         "SELECT MAX(id) FROM {namespace}_{identifier}.indexmetadataentity LIMIT 1"
     );
 
     let row = sqlx::query(&query).fetch_one(conn).await?;
-    let id: i32 = match row.try_get(0) {
+    let id: i64 = match row.try_get(0) {
         Ok(id) => id,
         Err(_e) => return Ok(1),
     };
 
-    Ok(id as u32)
+    Ok(id as u64)
 }
 
 // TODO: https://github.com/FuelLabs/fuel-indexer/issues/251
